refactor: extract shared matte helpers out of keying.py
